Add isUserLogin() to the inspection actions

+++ mpgk-react-frontend/src/actions/inspectionActions.js
@@ -10,16 +10,18 @@ import {
 	GET_CONNECTION
 } from "./types";
 import isUserLogin from "../securityUtils/isUserLogin";
-import {hostName} from "./host";
+import { hostName } from "./host";
 
 export const addInspection = (inspection, history) => async (dispatch) => {
 	try {
-		await axios.post(`${hostName}/api/inspections`, inspection);
-		history.push("/inspections");
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
+		if (isUserLogin()) {
+			await axios.post(`${hostName}/api/inspections`, inspection);
+			history.push("/inspections");
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -29,11 +31,13 @@ export const addInspection = (inspection, history) => async (dispatch) => {
 };
 
 export const getInspections = () => async (dispatch) => {
-	const res = await axios.get(`${hostName}/api/inspections`);
-	dispatch({
-		type: GET_INSPECTIONS,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get(`${hostName}/api/inspections`);
+		dispatch({
+			type: GET_INSPECTIONS,
+			payload: res.data
+		});
+	}
 };
 
 export const updateInspection = (inspectionId, updatedInspection) => async (
@@ -64,12 +68,14 @@ export const updateInspection = (inspectionId, updatedInspection) => async (
 };
 
 export const deleteInspection = (inspectionId) => async (dispatch) => {
-	if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie przeglądu")) {
-		await axios.delete(`${hostName}/api/inspections/${inspectionId}`);
-		dispatch({
-			type: DELETE_INSPECTION,
-			payload: inspectionId
-		});
+	if (isUserLogin()) {
+		if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie przeglądu")) {
+			await axios.delete(`${hostName}/api/inspections/${inspectionId}`);
+			dispatch({
+				type: DELETE_INSPECTION,
+				payload: inspectionId
+			});
+		}
 	}
 };
 
@@ -77,7 +83,7 @@ export const getInspectionByConnection = (connectionId, history) => async (
 	dispatch
 ) => {
 	try {
-		if (isUserLogin) {
+		if (isUserLogin()) {
 			const res = await axios.get(
 				`${hostName}/api/inspections/list/${connectionId}/execute`
 			);
@@ -97,13 +103,15 @@ export const getOverdueByConnection = (
 	history
 ) => async (dispatch) => {
 	try {
-		const res = await axios.get(
-			`${hostName}/api/inspections/list/${connectionId}/overdue/${endTime}`
-		);
-		dispatch({
-			type: GET_OVERDUE_BY_CONNECTION,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(
+				`${hostName}/api/inspections/list/${connectionId}/overdue/${endTime}`
+			);
+			dispatch({
+				type: GET_OVERDUE_BY_CONNECTION,
+				payload: res.data
+			});
+		}
 	} catch (error) {
 		history.push("/");
 	}
@@ -113,11 +121,15 @@ export const getInspectionsByConnection = (connectionId, history) => async (
 	dispatch
 ) => {
 	try {
-		const res = await axios.get(`${hostName}/api/inspections/list/${connectionId}`);
-		dispatch({
-			type: GET_CONNECTION,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(
+				`${hostName}/api/inspections/list/${connectionId}`
+			);
+			dispatch({
+				type: GET_CONNECTION,
+				payload: res.data
+			});
+		}
 	} catch (error) {
 		history.push("/");
 	}
@@ -125,11 +137,13 @@ export const getInspectionsByConnection = (connectionId, history) => async (
 
 export const getActionsByName = (history) => async (dispatch) => {
 	try {
-		const res = await axios.get(`${hostName}/api/inspections/nitrogen`);
-		dispatch({
-			type: GET_INSPECTIONS,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(`${hostName}/api/inspections/nitrogen`);
+			dispatch({
+				type: GET_INSPECTIONS,
+				payload: res.data
+			});
+		}
 	} catch (error) {
 		history.push("/");
 	}
@@ -143,15 +157,17 @@ export const getConnectionAndStartTimeBetween = (
 	history
 ) => async (dispatch) => {
 	try {
-		const res = await axios.get(
-			`${hostName}/api/inspections/list/${id}/from/${startTime}/to/${endTime}?type=${typeName}`
-		);
-		dispatch({
-			type: GET_CONNECTION,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(
+				`${hostName}/api/inspections/list/${id}/from/${startTime}/to/${endTime}?type=${typeName}`
+			);
+			dispatch({
+				type: GET_CONNECTION,
+				payload: res.data
+			});
 
-		return res;
+			return res;
+		}
 	} catch (error) {
 		history.push("/");
 	}
@@ -164,13 +180,15 @@ export const getInspectionByConnectionAndStartTimeAndEndTime = (
 	history
 ) => async (dispatch) => {
 	try {
-		const res = await axios.get(
-			`${hostName}/api/inspections/list/${connectionId}/${startTime}/to/${endTime}/show`
-		);
-		dispatch({
-			type: GET_INSPECTIONS,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(
+				`${hostName}/api/inspections/list/${connectionId}/${startTime}/to/${endTime}/show`
+			);
+			dispatch({
+				type: GET_INSPECTIONS,
+				payload: res.data
+			});
+		}
 	} catch (error) {
 		history.push("/");
 	}
